+++ phpstan.neon
@@ -71,6 +71,7 @@ parameters:
 
         # tests or coding standars?
         - '#Constructor of class [\s\S]+ has an unused parameter \$html#' # 2
+        - '#Call to an undefined method Nette\\Utils\\Html::add\(\)#' # 1
 
     excludes_analyse:
         # test files
