+++ tests/Rector/MethodCall/MethodNameReplacerRector/Correct/SomeClass.php
