+++ tests/Rector/MethodCall/MethodNameReplacerRector/Correct/correct.php.inc
