+++ tests/Rector/MethodCall/MethodNameReplacerRector/Correct/correct2.php.inc
