+++ tests/Rector/MethodCall/MethodNameReplacerRector/Correct/correct3.php.inc
