+++ tests/Rector/MethodCall/MethodNameReplacerRector/Correct/correct4.php.inc
