+++ tests/Rector/MethodCall/MethodNameReplacerRector/Correct/correct5.php.inc
