+++ tests/Rector/MethodCall/MethodNameReplacerRector/Correct/correct6.php.inc
