+++ tests/Rector/MethodCall/MethodNameReplacerRector/Correct/correct7.php.inc
