+++ tests/Rector/MethodCall/MethodNameReplacerRector/Wrong/SomeClass.php
