+++ tests/Rector/MethodCall/MethodNameReplacerRector/Wrong/wrong.php.inc
