+++ tests/Rector/MethodCall/MethodNameReplacerRector/Wrong/wrong2.php.inc
