+++ tests/Rector/MethodCall/MethodNameReplacerRector/Wrong/wrong3.php.inc
