+++ tests/Rector/MethodCall/MethodNameReplacerRector/Wrong/wrong4.php.inc
