+++ tests/Rector/MethodCall/MethodNameReplacerRector/Wrong/wrong5.php.inc
