+++ tests/Rector/MethodCall/MethodNameReplacerRector/Wrong/wrong6.php.inc
