+++ tests/Rector/MethodCall/MethodNameReplacerRector/Wrong/wrong7.php.inc
